+++ packages/orq-rc/src/lib/config.ts
@@ -68,7 +68,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "3.13.0-rc.18",
-  genVersion: "2.706.0",
-  userAgent: "speakeasy-sdk/typescript 3.13.0-rc.18 2.706.0 2.0 @orq-ai/node",
+  sdkVersion: "3.13.0-rc.19",
+  genVersion: "2.707.0",
+  userAgent: "speakeasy-sdk/typescript 3.13.0-rc.19 2.707.0 2.0 @orq-ai/node",
 } as const;